requirements.txt
python-dotenv
discord.py==2.2.3